+++ android/src/main/kotlin/com/godwin/calendar_events/CalenderError.kt
@@ -1,3 +1,3 @@
 package com.godwin.calendar_events
 
-class CalenderError(val errorCode: String,val details:String)
+class CalendarError(val errorCode: String,val details:String)