+++ android/src/main/kotlin/com/godwin/calendar_events/PermissionResultCallback.kt
@@ -2,5 +2,5 @@ package com.godwin.calendar_events
 
 interface PermissionResultCallback {
     fun onSuccess()
-    fun onFailed(error: CalenderError)
+    fun onFailed(error: CalendarError)
 }