+++ android/src/main/kotlin/com/godwin/calendar_events/Result.kt
@@ -1,7 +1,12 @@
 package com.godwin.calendar_events
 
-sealed interface CalenderListResult
+sealed interface CalendarListResult
 
-class CalenderListSuccess(val list: List<Map<String, Any?>>) : CalenderListResult
-class CalenderListFailed(val error: CalenderError) : CalenderListResult
+class CalendarListSuccess(val list: List<Map<String, Any?>>) : CalendarListResult
+class CalendarListFailed(val error: CalendarError) : CalendarListResult
 
+sealed interface CalendarEventResult
+
+class CalendarEventSuccess(val event: CalendarEvent) : CalendarEventResult
+class CalendarEventDeleteSuccess(val eventId: String) : CalendarEventResult
+class CalendarEventFailed(val error: CalendarError) : CalendarEventResult