+++ ios/Classes/CalendarEventsPlugin.swift
@@ -23,13 +23,61 @@ public class CalendarEventsPlugin: NSObject, FlutterPlugin {
         case "addEvent":
             if let map = call.arguments as? [String: Any] {
                 let calendarEvent =  CalendarUtil.convertToEvent(from: map)
-                let error = calendarManager.addEvent(calendarEvent: calendarEvent)
-                if(error != nil){
-                    result(FlutterError(code:error!.errorCode,message: error?.details,details: nil))
-                }else {
-                    result(1)
+                let eventresult: CalendarEventResult = calendarManager.addEvent(calendarEvent: calendarEvent)
+                switch eventresult{
+                case .success(let event) :
+                    result(["eventId": event.eventId])
+                    break
+                case .failed(let error) :
+                    result(FlutterError(code:error.errorCode,message: error.details,details: nil))
+                    break
+                case .deleteSuccess(eventId: _):
+                    break
+                }
+            } else {
+                result(FlutterError(
+                    code: "ARGUMENT_ERROR",
+                    message: "Invalid arguments",
+                    details: nil
+                ))
+            }
+            break
+        case "updateEvent":
+            if let map = call.arguments as? [String: Any] {
+                let calendarEvent =  CalendarUtil.convertToEvent(from: map)
+                let eventresult = calendarManager.updateEvent(calendarEvent: calendarEvent)
+                switch eventresult{
+                case .success(let event) :
+                    result(["eventId": event.eventId])
+                    break
+                case .failed(let error) :
+                    result(FlutterError(code:error.errorCode,message: error.details,details: nil))
+                    break
+                case .deleteSuccess(eventId: _):
+                    break
                 }
                 
+            } else {
+                result(FlutterError(
+                    code: "ARGUMENT_ERROR",
+                    message: "Invalid arguments",
+                    details: nil
+                ))
+            }
+            break
+        case "deleteEvent":
+            if let map = call.arguments as? [String: Any] {
+                let eventId = map["eventId"] as? String
+                let eventresult = calendarManager.deleteEvent(eventId: eventId)
+                switch eventresult{
+                case .success(_) :
+                    break
+                case .failed(let error) :
+                    result(FlutterError(code:error.errorCode,message: error.details,details: nil))
+                    break
+                case .deleteSuccess(eventId: let eventId):
+                    result(["eventId": eventId])
+                }
             } else {
                 result(FlutterError(
                     code: "ARGUMENT_ERROR",